swdptap: Cleaned up the naming and and signatures of the SWDP implementation functions

+++ src/platforms/hosted/dap.h
@@ -83,7 +83,7 @@ void dap_write_single(ADIv5_AP_t *ap, uint32_t dest, const void *src, enum align
 int dbg_dap_cmd(uint8_t *data, int size, int rsize);
 void dap_jtagtap_tdi_tdo_seq(uint8_t *data_out, bool final_tms, const uint8_t *tms, const uint8_t *data_in, size_t clock_cycles);
 int dap_jtag_configure(void);
-void dap_swdptap_seq_out(uint32_t tms_states, int clock_cycles);
-void dap_swdptap_seq_out_parity(uint32_t tms_states, int clock_cycles);
+void dap_swdptap_seq_out(uint32_t tms_states, size_t clock_cycles);
+void dap_swdptap_seq_out_parity(uint32_t tms_states, size_t clock_cycles);
 
 #endif // _DAP_H_
